requirements.txt
# Web UI
# streamlit==1.28.0
streamlit==1.35.0
# Data handling
pandas==2.0.3

# Image processing
Pillow==10.0.0
pdf2image==1.16.3

# Google Cloud APIs
google-cloud-documentai==2.18.0
google-cloud-storage==2.14.0

# Core protocol support
protobuf==4.25.1

# Table formatting (used in app.py)
tabulate==0.9.0
